refactoring_load_lights
lightsの読み込みをリファクタリング
